refactor: extract agent-rules into cmd/agent_rules.rs
Move AgentMode, AgentPlatform, AgentRulesArgs, AGENT_RULES_GENERATED_MARKER,
generate_agent_rules(), run(), and all 11 tests from cmd/mod.rs into a
dedicated cmd/agent_rules.rs module.

cmd/mod.rs drops from ~740 lines to ~180 lines, containing only the
Command enum, dispatch(), and project config loading.

Updates cmd/init.rs to import directly from cmd::agent_rules instead of
relying on re-exports through cmd::mod.
